\name{a_vs_R0dN}
\alias{a_vs_R0dN}
\docType{data}
\title{A pre-calibrated spline function needed to translate the number of upper/lower records in to signa-to-noise ratios.
}
\description{
This function is needed by There should not be any need to use directly this function.}

\seealso{estimateSNR}
\keyword{datasets}
