\name{f}
\alias{f}
\docType{data}
\title{A pre-calibrated spline function needed to translate the number of upper/lower records in to signa-to-noise ratios.
}
\description{This function is needed to translate the number of records to Sharpe ratios. There should not be any need to use directly this function.}

\seealso{estimateSNR}
\keyword{datasets}
