% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{num_records_down}
\alias{num_records_down}
\title{Computes the number of lower records of the cumulative sum of \code{x}}
\usage{
num_records_down(x)
}
\arguments{
\item{x}{a vector of sample values}
}
\value{
the number of lower records
}
\description{
Computes the number of lower records of the cumulative sum of \code{x}
}
