% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{num_records_up}
\alias{num_records_up}
\title{Computes the number of upper records of the cumulative sum of \code{x}}
\usage{
num_records_up(x)
}
\arguments{
\item{x}{a vector of sample values}
}
\value{
the number of upper records
}
\description{
Computes the number of upper records of the cumulative sum of \code{x}
}
